+++ tests/input/variable_annotations.gd
@@ -9,3 +9,31 @@ var jump_time_to_peak:=0.37
 
 @export_group("my group")
 @export var v = 1
+
+@warning_ignore("unsafe_property_access")
+var url: String = (
+	_js_window.location.hash.trim_prefix("#").trim_prefix("/")
+)
+
+# @export and @onready annotations should merge with the variable declaration
+# even when the expression wraps over multiple lines.
+@export
+var exported_url: String = (
+	_js_window.location.hash.trim_prefix("#").trim_prefix("/") + _js_window.location.hash.trim_prefix("#").trim_prefix("/")
+)
+
+@onready
+var onready_url: String = (
+	_js_window.location.hash.trim_prefix("#").trim_prefix("/") + _js_window.location.hash.trim_prefix("#").trim_prefix("/")
+)
+
+class AnnotationsInClassBody:
+	@warning_ignore("unsafe_property_access")
+	var url: String = (
+		_js_window.location.hash.trim_prefix("#").trim_prefix("/")
+	)
+
+	@export
+	var exported_url: String = (
+		_js_window.location.hash.trim_prefix("#").trim_prefix("/") + _js_window.location.hash.trim_prefix("#").trim_prefix("/")
+	)